requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.3.2               # via kombu
billiard==3.5.0.3         # via celery
celery[redis]==4.1.1
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
dj-database-url==0.5.0    # via django-heroku
django-bootstrap4==0.0.6
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-heroku==0.3.1
django-meta==1.4.1
django-rosetta==0.8.1
django==2.1
gunicorn==19.8.1
idna==2.6                 # via requests
kombu==4.2.1              # via celery
microsofttranslator==0.8  # via django-rosetta
pillow==5.1.0
polib==1.1.0              # via django-rosetta
psycopg2==2.7.4
python-decouple==3.1
python-memcached==1.59
pytz==2018.4              # via celery, django
redis==2.10.6             # via celery
requests==2.18.4
six==1.11.0               # via django-rosetta, microsofttranslator, python-memcached
urllib3==1.22             # via requests
vine==1.1.4               # via amqp
whitenoise==3.3.1
